requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-bootstrap5==25.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
psycopg==3.2.5
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.9.0
